refactor(nutricion): generaliza los tabs de evaluación por endpoint
Extrae un genérico EndpointEvalTab parametrizado por config (mata la
triplicación de BioquimicaTab/NutricionalTab/ExamFisicaTab) y mueve cada
tab a su propio archivo, junto con el MonitoreoTab de sueño/act. física.
PatientHistoriaNutricion queda como agregador delgado y deja de importar
todos los forms/hooks de golpe; cada tab trae sus imports y su fetch se
dispara solo al montarse.
